+++ congresso/__init__.py
@@ -6,6 +6,6 @@ from .camara import Camara
 try:
     __version__ = pkg_resources.get_distribution(__name__).version
 except:
-    __version__ = 'unknown'
+    __version__ = '0.1.0 '