+++ congresso/camara/__init__.py
@@ -9,16 +9,16 @@ class Camara(object):
 
     The instance has attributes ``deputados``, ``orgaos``, ``proposicoes``, ``sessoesreunioes``,
     that provide access to instances of
-    :class:`~cn_api_client.camara.DeputadosClient`,
-    :class:`~cn_api_client.camara.OrgaosClient`,
-    :class:`~cn_api_client.camara.ProposicoesClient` and
-    :class:`~cn_api_client.camara.SessoesClient` respectively. This is the
+    :class:`~congresso.camara.DeputadosClient`,
+    :class:`~congresso.camara.OrgaosClient`,
+    :class:`~congresso.camara.ProposicoesClient` and
+    :class:`~congresso.camara.SessoesClient` respectively. This is the
     preferred (and only supported) way to get access to those classes and their
     methods.
 
     You can initialize your connection class by:
 
-        camara = cn_api_client.Camara()
+        camara = congresso.Camara()
 
     and you'll be ready to use the API on your Python projetct.
     """