+++ congresso/senado/senado.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """
-This is a set of cn_api_client intended to hook the API from the Brazilian Senate.
+This is a set of congresso intended to hook the API from the Brazilian Senate.
 
  The API documentation can be red at http://legis.senado.leg.br/dadosabertos/docs/
 """