+++ CHANGES.rst
@@ -5,6 +5,6 @@ Changelog
 Version 0.1
 ===========
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+- Low Level API Camara dos Deputados Client built
+- Tests for connection and utils
+- Very basic connection handling